+++ apps/dine-pleiepenger/src/server/api-models/SakSchema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import { parseMaybeDateStringToDate } from '../../utils/jsonParseUtils';
+
+export type Sak = z.infer<typeof SakSchema>;
+
+export const SakSchema = z.object({
+    saksbehandlingsFrist: z.union([z.preprocess((val) => parseMaybeDateStringToDate(val), z.date()), z.undefined()]),
+});