+++ apps/dine-pleiepenger/src/server/api-models/SaksbehandlingstidSchema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export type Saksbehandlingstid = z.infer<typeof SaksbehandlingstidSchema>;
+
+export const SaksbehandlingstidSchema = z.object({
+    saksbehandlingstidUker: z.number(),
+});