+++ apps/dine-pleiepenger/src/types/InnsynData.ts
@@ -1,13 +1,13 @@
 import { Søker } from '../server/api-models/SøkerSchema';
 import { Mellomlagringer } from './Mellomlagring';
+import { Sak } from './Sak';
 import { Søknad } from './Søknad';
 
 export interface Innsynsdata {
     søker: Søker;
     søknader: Søknad[];
     mellomlagring: Mellomlagringer;
-    svarfrist?: Date;
-    behandlingstid?: {
-        uker: number;
-    };
+    saker: Sak[];
+    saksbehandlingstidUker?: number;
+    harSak: boolean;
 }