+++ apps/dine-pleiepenger/src/utils/env.ts
@@ -5,6 +5,7 @@ export const publicEnvSchema = z.object({
     NEXT_PUBLIC_BASE_PATH: z.string(),
     NEXT_PUBLIC_LOGIN_URL: z.string(),
     NEXT_PUBLIC_RUNTIME_ENVIRONMENT: z.union([z.string(), z.undefined()]),
+    NEXT_PUBLIC_API_URL_K9_SAK_INNSYN: z.union([z.string(), z.undefined()]),
     NEXT_PUBLIC_API_URL_INNSYN: z.union([z.string(), z.undefined()]),
     NEXT_PUBLIC_API_URL_BRUKERDIALOG: z.union([z.string(), z.undefined()]),
 
@@ -21,7 +22,7 @@ export const publicEnvSchema = z.object({
     NEXT_PUBLIC_MIN_SIDE_URL: z.string(),
     NEXT_PUBLIC_PLEIEPENGER_URL: z.string(),
     NEXT_PUBLIC_REGELVERK_INFO_URL: z.string(),
-    NEXT_PUBLIC_SAKBEHANDLINGSTID_INFO_URL: z.string(),
+    NEXT_PUBLIC_SAKSBEHANDLINGSTID_INFO_URL: z.string(),
     NEXT_PUBLIC_SAKSOVERSIKT: z.string(),
     NEXT_PUBLIC_SKRIV_TIL_OSS: z.string(),
     NEXT_PUBLIC_SYKDOM_I_FAMILIEN_INFO_URL: z.string(),
@@ -30,14 +31,15 @@ export const publicEnvSchema = z.object({
     NEXT_PUBLIC_UTBETALINGSOVERSIKT_URL: z.string(),
 
     /** Features */
-    NEXT_PUBLIC_FEATURE_SVARFRIST: z.union([z.literal('on'), z.literal('off'), z.undefined()]),
-    NEXT_PUBLIC_FEATURE_BEHANDLINGSTID: z.union([z.literal('on'), z.literal('off'), z.undefined()]),
+    NEXT_PUBLIC_FEATURE_HENT_SAKER: z.union([z.literal('on'), z.literal('off'), z.undefined()]),
+    NEXT_PUBLIC_FEATURE_HENT_BEHANDLINGSTID: z.union([z.literal('on'), z.literal('off'), z.undefined()]),
 });
 
 export type ServerEnv = z.infer<typeof serverEnvSchema>;
 export const serverEnvSchema = z.object({
     NEXT_PUBLIC_INNSYN_BACKEND_SCOPE: z.string(),
     NEXT_PUBLIC_BRUKERDIALOG_BACKEND_SCOPE: z.string(),
+    NEXT_PUBLIC_K9_SAK_INNSYN_BACKEND_SCOPE: z.string(),
     IDPORTEN_CLIENT_ID: z.union([z.string(), z.undefined()]),
     IDPORTEN_WELL_KNOWN_URL: z.union([z.string(), z.undefined()]),
     TOKEN_X_WELL_KNOWN_URL: z.union([z.string(), z.undefined()]),
@@ -58,8 +60,9 @@ export const browserEnv = publicEnvSchema.parse({
     NEXT_PUBLIC_APPSTATUS_PROJECT_ID: process.env.NEXT_PUBLIC_APPSTATUS_PROJECT_ID,
     NEXT_PUBLIC_APPSTATUS_DATASET: process.env.NEXT_PUBLIC_APPSTATUS_DATASET,
     NEXT_PUBLIC_API_URL_BRUKERDIALOG: process.env.NEXT_PUBLIC_API_URL_BRUKERDIALOG,
+    NEXT_PUBLIC_API_URL_K9_SAK_INNSYN: process.env.NEXT_PUBLIC_API_URL_K9_SAK_INNSYN,
     NEXT_PUBLIC_API_URL_INNSYN: process.env.NEXT_PUBLIC_API_URL_INNSYN,
-    NEXT_PUBLIC_SAKBEHANDLINGSTID_INFO_URL: process.env.NEXT_PUBLIC_SAKBEHANDLINGSTID_INFO_URL,
+    NEXT_PUBLIC_SAKSBEHANDLINGSTID_INFO_URL: process.env.NEXT_PUBLIC_SAKSBEHANDLINGSTID_INFO_URL,
     NEXT_PUBLIC_SYKDOM_I_FAMILIEN_INFO_URL: process.env.NEXT_PUBLIC_SYKDOM_I_FAMILIEN_INFO_URL,
     NEXT_PUBLIC_ETTERSENDELSE_PLEIEPENGER_URL: process.env.NEXT_PUBLIC_ETTERSENDELSE_PLEIEPENGER_URL,
     NEXT_PUBLIC_KLAGE_INFO_URL: process.env.NEXT_PUBLIC_KLAGE_INFO_URL,
@@ -74,13 +77,14 @@ export const browserEnv = publicEnvSchema.parse({
     NEXT_PUBLIC_ARBEIDSGIVER_PLEIEPENGER_URL: process.env.NEXT_PUBLIC_ARBEIDSGIVER_PLEIEPENGER_URL,
     NEXT_PUBLIC_MINSIDE_DOKUMENTOVERSIKT_URL: process.env.NEXT_PUBLIC_MINSIDE_DOKUMENTOVERSIKT_URL,
     NEXT_PUBLIC_UTBETALINGSOVERSIKT_URL: process.env.NEXT_PUBLIC_UTBETALINGSOVERSIKT_URL,
-    NEXT_PUBLIC_FEATURE_SVARFRIST: process.env.NEXT_PUBLIC_FEATURE_SVARFRIST,
-    NEXT_PUBLIC_FEATURE_BEHANDLINGSTID: process.env.NEXT_PUBLIC_FEATURE_BEHANDLINGSTID,
+    NEXT_PUBLIC_FEATURE_HENT_SAKER: process.env.NEXT_PUBLIC_FEATURE_HENT_SAKER,
+    NEXT_PUBLIC_FEATURE_HENT_BEHANDLINGSTID: process.env.NEXT_PUBLIC_FEATURE_HENT_BEHANDLINGSTID,
 } satisfies Record<keyof PublicEnv, string | undefined>);
 
 const getRawServerConfig = (): Partial<unknown> =>
     ({
         // Provided by nais-*.yml
+        NEXT_PUBLIC_K9_SAK_INNSYN_BACKEND_SCOPE: process.env.NEXT_PUBLIC_K9_SAK_INNSYN_BACKEND_SCOPE,
         NEXT_PUBLIC_INNSYN_BACKEND_SCOPE: process.env.NEXT_PUBLIC_INNSYN_BACKEND_SCOPE,
         NEXT_PUBLIC_BRUKERDIALOG_BACKEND_SCOPE: process.env.NEXT_PUBLIC_BRUKERDIALOG_BACKEND_SCOPE,