+++ apps/dine-pleiepenger/src/utils/features.ts
@@ -1,6 +1,6 @@
 import { browserEnv } from './env';
 
 export const Feature = {
-    HENT_SVARFRIST: browserEnv.NEXT_PUBLIC_FEATURE_SVARFRIST === 'on',
-    HENT_BEHANDLINGSTID: browserEnv.NEXT_PUBLIC_FEATURE_BEHANDLINGSTID === 'on',
+    HENT_SAKER: browserEnv.NEXT_PUBLIC_FEATURE_HENT_SAKER === 'on',
+    HENT_BEHANDLINGSTID: browserEnv.NEXT_PUBLIC_FEATURE_HENT_BEHANDLINGSTID === 'on',
 };