+++ apps/omsorgsdager-kalkulator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/omsorgsdager-kalkulator
 
+## 3.5.18
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor og patch
+
+## 3.5.17
+
+### Patch Changes
+
+- Pakkeoppdateringer patch/minor
+
 ## 3.5.16
 
 ### Patch Changes