+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/appstatus-react-ds
 
+## 2.17.21
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor og patch
+
+## 2.17.20
+
+### Patch Changes
+
+- Pakkeoppdateringer patch/minor
+
 ## 2.17.19
 
 ### Patch Changes