+++ packages/sif-common-formik-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-formik-ds
 
+## 1.7.12
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor og patch
+
+## 1.7.11
+
+### Patch Changes
+
+- Pakkeoppdateringer patch/minor
+
 ## 1.7.10
 
 ### Patch Changes