+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/sif-common-forms-ds
 
+## 8.19.1
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor og patch
+
+## 8.19.0
+
+### Minor Changes
+
+- Legge til rette for nynorsk i annetBarn dialog
+
+## 8.18.3
+
+### Patch Changes
+
+- Pakkeoppdateringer patch/minor
+
 ## 8.18.2
 
 ### Patch Changes